+++ composer.json
@@ -27,5 +27,8 @@
         "platform": {
             "php": "8.2.0"
         }
-    }
+    },
+    "bin": [
+
+    ]
 }
