+++ tests/RouterTest.php
@@ -5,8 +5,6 @@ declare(strict_types=1);
 namespace willitscale\StreetlampTests;
 
 use willitscale\Streetlamp\Enums\MediaType;
-use willitscale\Streetlamp\Exceptions\InvalidParameterTypeException;
-use willitscale\Streetlamp\Exceptions\Validators\InvalidParameterFailedToPassFilterValidation;
 use willitscale\StreetlampTest\RouteTestCase;
 
 class RouterTest extends RouteTestCase
@@ -120,174 +118,5 @@ class RouterTest extends RouteTestCase
         );
         $response = $router->route(true);
         $this->assertEquals($data, $response);
-    }
-
-    public function testRouterGetMethodWithPathParameterThatValidatesInput(): void
-    {
-        $expectedResponse = 99;
-
-        $router = $this->setupRouter(
-            'GET',
-            '/validator/' . $expectedResponse,
-            MediaType::TEXT_HTML->value,
-            __DIR__,
-            self::COMPOSER_TEST_FILE
-        );
-        $response = $router->route(true);
-        $this->assertEquals($expectedResponse, $response);
-    }
-
-    public function testRouterCacheAlwaysReturnsTheInitialCachedValue(): void
-    {
-        $expectedCacheValue = 99;
-        $unexpectedCacheValue = 23;
-
-        $router = $this->setupRouter(
-            'GET',
-            '/cache/' . $expectedCacheValue,
-            MediaType::TEXT_HTML->value,
-            __DIR__,
-            self::COMPOSER_TEST_FILE
-        );
-
-        $router->route(true);
-
-        $router = $this->setupRouter(
-            'GET',
-            '/cache/' . $unexpectedCacheValue,
-            MediaType::TEXT_HTML->value,
-            __DIR__,
-            self::COMPOSER_TEST_FILE
-        );
-
-        $response = $router->route(true);
-
-        $this->assertEquals($expectedCacheValue, $response);
-    }
-
-    public function testRouterCacheAlwaysReturnsTheParameterCachedValue(): void
-    {
-        $firstCachedValue = 99;
-        $secondCachedValue = 23;
-
-        $router = $this->setupRouter(
-            'GET',
-            '/cache/parameter/' . $firstCachedValue,
-            MediaType::TEXT_HTML->value,
-            __DIR__,
-            self::COMPOSER_TEST_FILE
-        );
-
-        $response = $router->route(true);
-
-        $this->assertEquals($firstCachedValue, $response);
-
-        $router = $this->setupRouter(
-            'GET',
-            '/cache/parameter/' . $secondCachedValue,
-            MediaType::TEXT_HTML->value,
-            __DIR__,
-            self::COMPOSER_TEST_FILE
-        );
-
-        $response = $router->route(true);
-
-        $this->assertEquals($secondCachedValue, $response);
-    }
-
-    public function testRouterDataMappingCorrectlyCreatesAnObject(): void
-    {
-        $testData = [
-            'name' => 'Test',
-            'age' => 123
-        ];
-
-        file_put_contents(self::TEST_BODY_FILE, json_encode($testData));
-
-        $router = $this->setupRouter(
-            'POST',
-            '/data/validation',
-            MediaType::APPLICATION_JSON->value,
-            __DIR__,
-            self::COMPOSER_TEST_FILE
-        );
-
-        $response = $router->route(true);
-
-        $this->assertEquals($testData, json_decode($response, true));
-    }
-
-    public function testRouterDataMappingWithIncorrectDataFailsToCreateObject(): void
-    {
-        $testData = [
-            'name' => 'Test'
-        ];
-
-        $testFile =
-            file_put_contents(self::TEST_BODY_FILE, json_encode($testData));
-
-        $router = $this->setupRouter(
-            'POST',
-            '/data/validation',
-            MediaType::APPLICATION_JSON->value,
-            __DIR__,
-            self::COMPOSER_TEST_FILE
-        );
-
-        $this->expectException(InvalidParameterTypeException::class);
-        $router->route(true);
-    }
-
-    public function testRouterDataMappingCorrectlyCreatesAnArrayOfObjects(): void
-    {
-        $testData = [
-            [
-                'name' => 'Test',
-                'age' => 123
-            ],
-            [
-                'name' => 'Tester',
-                'age' => 456
-            ]
-        ];
-
-        file_put_contents(self::TEST_BODY_FILE, json_encode($testData));
-
-        $router = $this->setupRouter(
-            'POST',
-            '/data/validations',
-            MediaType::APPLICATION_JSON->value,
-            __DIR__,
-            self::COMPOSER_TEST_FILE
-        );
-
-        $response = $router->route(true);
-        $this->assertEquals($testData, json_decode($response, true));
-    }
-
-    public function testRouterDataMappingWithIncorrectDataFailsToCreateAnArrayOfObjects(): void
-    {
-        $testData = [
-            [
-                'name' => 'Test',
-                'age' => 123
-            ],
-            [
-                'name' => 'Tester'
-            ]
-        ];
-
-        file_put_contents(self::TEST_BODY_FILE, json_encode($testData));
-
-        $router = $this->setupRouter(
-            'POST',
-            '/data/validations',
-            MediaType::APPLICATION_JSON->value,
-            __DIR__,
-            self::COMPOSER_TEST_FILE
-        );
-
-        $this->expectException(InvalidParameterFailedToPassFilterValidation::class);
-        $router->route(true);
     }
 }
